refactor(web): extract the marquee, snap-guide and edge-highlight overlays into components
Three inline canvas-space overlay SVGs leave SpatialEditor's JSX as sibling
components of SelectionOverlay/GhostOverlay, verbatim: MarqueeOverlay,
SnapGuidesOverlay (which takes the guideSpan memo with it — the span was
computed for nothing else), and EdgeSelectionHighlight (which keeps the
resolve-in-render null gate for a deleted edge). All three join
manipulation-tokens.test.ts's scanned overlay set (5 -> 8 files), so the
no-hex-literal rule keeps covering the moved markup.

Behavior-preserving: marquee, snap, edge-selection-coherence,
edge-select-delete, selection-draw and live-drag browser suites (29 tests)
and the spatial-editor jsdom suite (45 files / 494 tests) pass unchanged.
SpatialEditor.tsx: 2622 -> 2491 lines (3364 at the start of this split).
